them cac goi dv, thanh toan thanh cong, kiem tra info thanh toan toan bo user

+++ src/main/java/movieapp/webmovie/dto/UserResponseDTO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class UserResponseDTO {
-    private Integer userID;
+    private Long userID;
     private String name;
     private String email;
     private String phone;
